# docs/api-audit/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tecbunny API Surface",
    "version": "1.0.0",
    "description": "Generated from repository source by scripts/audit-api-surface.mjs."
  },
  "servers": [
    {
      "url": "{{API_BASE_URL}}"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/messages/media": {
      "post": {
        "description": "POST /api/messages/media",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Observed or expected 400 response"
          },
          "403": {
            "description": "Observed or expected 403 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-route-file": "apps/waba/src/app/api/messages/media/route.ts",
        "x-database-tables": [
          "whatsapp_media"
        ],
        "x-frontend-usage": [
          "apps/waba/src/app/page.tsx"
        ]
      }
    },
...
    "/api/superadmin/inquiries/{id}/assignment": {
      "patch": {
        "description": "PATCH /api/superadmin/inquiries/{id}/assignment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Observed or expected 400 response"
          },
          "403": {
            "description": "Observed or expected 403 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-route-file": "apps/superadmin/src/app/api/superadmin/inquiries/[id]/assignment/route.ts",
        "x-database-tables": [
          "contact_messages",
          "profiles"
        ],
        "x-frontend-usage": []
      }
    },
    "/api/superadmin/services/ai-generate": {
      "post": {
        "description": "POST /api/superadmin/services/ai-generate",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Observed or expected 400 response"
          },
          "403": {
            "description": "Observed or expected 403 response"
          },
          "500": {
            "description": "Observed or expected 500 response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-route-file": "apps/superadmin/src/app/api/superadmin/services/ai-generate/route.ts",
        "x-database-tables": [],
        "x-frontend-usage": [
          "apps/superadmin/src/app/superadmin/mgmt/services/page.tsx"
        ]
      }
    },
...
  }
}